More crds.jwst.pipeline refactoring: Factored reference file generation code out of runtime pipeline module. Switched to using default YAML taken from prototype reference file.
