+++ crds/jwst/pipeline.py
@@ -15,16 +15,22 @@ things like "reference types used by a pipeline."
 
 """
 
+# --------------------------------------------------------------------------------------
+
+import os.path
 import fnmatch
 import re
 
 # import yaml
 
-import crds
-from crds.core import exceptions, log, utils, python23
-from crds.core.log import srepr
+# --------------------------------------------------------------------------------------
 
-# from jwst.stpipe import cmdline
+import crds
+from crds.core import exceptions, log, utils
+from crds.core.log import srepr
+from crds.client import api
+
+# --------------------------------------------------------------------------------------
 
 def test_header(calver, exp_type):
     header = {
@@ -35,180 +41,12 @@ def test_header(calver, exp_type):
         }
     return header
 
-'''
-DARK, (LED, LAMP, FLAT)  all level2a only
-DARK has special calwebb_dark
-
-All otheres go through both level2a and level2b
-'''
-
-HEADER_YAML = """
-author: CRDS
-descrip: "Reference used to determine pipeline configuration from dataset parameters."
-history: "First version generated from calcode .cfg files and EXP_TYPE/LEVEL mapping."
-instrument: SYSTEM
-pedigree: DUMMY
-reftype: CRDSCFG
-telescope: JWST
-useafter: 1900-01-01T00:00:00
-"""
-
-# Input manually
-# Order is important since the first pattern matching an exp_type in any given level wins.
-PIPELINE_CFGS_YAML = """
-pipeline_cfgs: [calwebb_dark.cfg, calwebb_sloper.cfg, calwebb_spec2.cfg, calniriss_soss2.cfg,
-  calwebb_image2.cfg]
-"""
-
-# Input manually
-LEVEL_PIPELINE_EXPTYPE_YAML = """
-level_pipeline_exptypes:
-    level2a:
-        - calwebb_dark.cfg: [FGS_DARK, MIR_DARK, NRC_DARK, NIS_DARK, NRS_DARK]
-
-        - calwebb_sloper.cfg: ["*"]
-
-    level2b:
-        - calwebb_spec2.cfg: [MIR_LRS-FIXEDSLIT, MIR_LRS-SLITLESS, MIR_MRS, NRS_FIXEDSLIT, 
-                             NRS_MSASPEC, NRS_IFU, NRS_BRIGHTOBJ, NRS_AUTOWAVE]
-
-        - calniriss_soss2.cfg: [NIS_SOSS]
-
-        - calwebb_image2.cfg: [NRC_IMAGE, NRC_TACQ, NRC_CORON, NRC_FOCUS, 
-                             MIR_IMAGE, MIR_TACQ, MIR_LYOT, MIR_4QPM, MIR_CORONCAL,
-                             NIS_IMAGE, NIS_AMI, NIS_TACQ,
-                             NRS_IMAGE, NRS_FOCUS, NRS_MIMF, NRS_BOTA, NRS_TACQ, NRS_TASLIT, NRS_TACONFIRM, NRS_CONFIRM,
-                             FGS_IMAGE, FGS_FOCUS]
-
-        - skip_2b.cfg: ["*DARK*", "*FLAT*", "*LED*", "*LAMP*", NIS_FOCUS, NIS_WFSS, NRS_AUTOWAVE]
-"""
-
-STEPS_TO_REFTYPE_EXCEPTIONS_YAML = """
-steps_to_reftypes_exceptions:
-    flat_field:
-        - case1:
-            exp_types: [NRS_FIXEDSLIT, NRS_IFU, NRS_MSASPEC]
-            reftypes: [dflat, fflat, sflat]
-        - case2:
-            exp_types: ["NRS_*"]
-            reftypes: []
-        - case3:
-            exp_types: ["*"]
-            reftypes: [flat]
-"""
 
 # --------------------------------------------------------------------------------------
 
-def generate_crdscfg_yaml():
-    """Generate the SYSTEM CRDSCFG reference YAML."""
-    import yaml
-    pipeline_cfgs = yaml.load(PIPELINE_CFGS_YAML)["pipeline_cfgs"]
-    pipeline_cfgs_to_steps, all_steps_to_reftypes = generate_pipeline_info(pipeline_cfgs)
-    crdscfg = HEADER_YAML + PIPELINE_CFGS_YAML + LEVEL_PIPELINE_EXPTYPE_YAML + "\n"
-    crdscfg += yaml.dump({"pipeline_cfgs_to_steps" : pipeline_cfgs_to_steps}) + "\n"
-    crdscfg += yaml.dump({"steps_to_reftypes" : all_steps_to_reftypes})
-    crdscfg += STEPS_TO_REFTYPE_EXCEPTIONS_YAML
-    return crdscfg
-    
-def generate_pipeline_info(pipeline_cfgs):
-    from jwst.stpipe import cmdline
-    pipeline_cfgs_to_steps = {}
-    all_steps_to_reftypes = {}
-    pipeline_cfgs_to_steps["skip_2b.cfg"] = []
-    for pipeline_cfg in pipeline_cfgs:
-        steps_to_reftypes = cmdline.steps_to_reftypes_from_config(pipeline_cfg)
-        pipeline_cfgs_to_steps[pipeline_cfg] = sorted(list(steps_to_reftypes.keys()))
-        all_steps_to_reftypes.update(steps_to_reftypes)
-    return pipeline_cfgs_to_steps, all_steps_to_reftypes
+HERE = os.path.dirname(__file__) or "."
 
-# --------------------------------------------------------------------------------------
-CRDSCFG_REFERENCE_YAML = '''
-author: CRDS
-descrip: "Reference used to determine pipeline configuration from dataset parameters."
-history: "First version generated from calcode .cfg files and EXP_TYPE/LEVEL mapping."
-instrument: SYSTEM
-pedigree: DUMMY
-reftype: CRDSCFG
-telescope: JWST
-useafter: 1900-01-01T00:00:00
-
-pipeline_cfgs: [calwebb_dark.cfg, calwebb_sloper.cfg, calwebb_spec2.cfg, calniriss_soss2.cfg,
-  calwebb_image2.cfg]
-
-level_pipeline_exptypes:
-    level2a:
-        - calwebb_dark.cfg: [FGS_DARK, MIR_DARK, NRC_DARK, NIS_DARK, NRS_DARK]
-
-        - calwebb_sloper.cfg: ["*"]
-
-    level2b:
-        - calwebb_spec2.cfg: [MIR_LRS-FIXEDSLIT, MIR_LRS-SLITLESS, MIR_MRS, NRS_FIXEDSLIT, 
-                             NRS_MSASPEC, NRS_IFU, NRS_BRIGHTOBJ, NRS_AUTOWAVE]
-
-        - calniriss_soss2.cfg: [NIS_SOSS]
-
-        - calwebb_image2.cfg: [NRC_IMAGE, NRC_TACQ, NRC_CORON, NRC_FOCUS, 
-                             MIR_IMAGE, MIR_TACQ, MIR_LYOT, MIR_4QPM, MIR_CORONCAL,
-                             NIS_IMAGE, NIS_AMI, NIS_TACQ,
-                             NRS_IMAGE, NRS_FOCUS, NRS_MIMF, NRS_BOTA, NRS_TACQ, NRS_TASLIT, NRS_TACONFIRM, NRS_CONFIRM,
-                             FGS_IMAGE, FGS_FOCUS]
-
-        - skip_2b.cfg: ["*DARK*", "*FLAT*", "*LED*", "*LAMP*", NIS_FOCUS, NIS_WFSS, NRS_AUTOWAVE]
-
-pipeline_cfgs_to_steps:
-  calniriss_soss2.cfg: [assign_wcs, bkg_subtract, cube_build, extract_1d, extract_2d,
-    flat_field, fringe, imprint_subtract, pathloss, photom, resample_spec, srctype,
-    straylight]
-  calwebb_dark.cfg: [dq_init, ipc, lastframe, linearity, refpix, rscd, saturation,
-    superbias]
-  calwebb_image2.cfg: [assign_wcs, flat_field, photom]
-  calwebb_sloper.cfg: [dark_current, dq_init, ipc, jump, lastframe, linearity, persistence,
-    ramp_fit, refpix, rscd, saturation, superbias]
-  calwebb_spec2.cfg: [assign_wcs, bkg_subtract, cube_build, extract_1d, extract_2d,
-    flat_field, fringe, imprint_subtract, pathloss, photom, resample_spec, srctype,
-    straylight]
-  skip_2b.cfg: []
-
-steps_to_reftypes:
-  assign_wcs: [camera, collimator, disperser, distortion, filteroffset, fore, fpa,
-    ifufore, ifupost, ifuslicer, msa, ote, regions, specwcs, v2v3, wavelengthrange]
-  bkg_subtract: []
-  cube_build: []
-  dark_current: [dark]
-  dq_init: [mask]
-  extract_1d: [extract1d]
-  extract_2d: []
-  flat_field: []
-  fringe: [fringe]
-  imprint_subtract: []
-  ipc: [ipc]
-  jump: [gain, readnoise]
-  lastframe: []
-  linearity: [linearity]
-  pathloss: [pathloss]
-  persistence: []
-  photom: [area, photom]
-  ramp_fit: [gain, readnoise]
-  refpix: [refpix]
-  resample_spec: [drizpars]
-  rscd: [rscd]
-  saturation: [saturation]
-  srctype: []
-  straylight: [straymask]
-  superbias: [superbias]
-
-steps_to_reftypes_exceptions:
-    flat_field:
-        - case1:
-            exp_types: [NRS_FIXEDSLIT, NRS_IFU, NRS_MSASPEC]
-            reftypes: [dflat, fflat, sflat]
-        - case2:
-            exp_types: ["NRS_*"]
-            reftypes: []
-        - case3:
-            exp_types: ["*"]
-            reftypes: [flat]
-'''
+DEFAULT_SYSTEM_CRDSCFG_PATH = os.path.join(HERE, "jwst_system_crdscfg.yaml")
 
 # --------------------------------------------------------------------------------------
 
@@ -232,8 +70,14 @@ def get_config_manager(context, cal_ver):
     and calibration s/w version `calver`.
     """
     import yaml
+    # default enables running on contexts that have no system crdscfg 
+    # reference, B7 and earlier.
+    refpath = DEFAULT_SYSTEM_CRDSCFG_PATH
     with log.verbose_warning_on_exception(
-            "Failed to load SYSTEM CRDSCFG for determining required reftypes.   Using build-7 default."):
+            "Failed locating SYSTEM CRDSCFG reference",
+            "under context", repr(context),
+            "and cal_ver", repr(cal_ver) + ".",
+            "Using build-7 default."):
         header = {
             "META.INSTRUMENT.NAME" : "SYSTEM", 
             "META.CALIBRATION_SOFTWARE_VERSION": cal_ver 
@@ -242,11 +86,10 @@ def get_config_manager(context, cal_ver):
         imap = pmap.get_imap("system")
         rmapping = imap.get_rmap("crdscfg")
         refpath = rmapping.get_bestref(header)
-        with open(refpath) as opened:
-            crdscfg =  yaml.load(opened)
-        return CrdsCfgManager(context, refpath, crdscfg)
-    crdscfg = yaml.load(python23.StringIO(CRDSCFG_REFERENCE_YAML))
-    return CrdsCfgManager(context, "no_system_crdscfg_reference_found", crdscfg)
+        api.dump_references(context, [refpath])
+    with open(refpath) as opened:
+        crdscfg =  yaml.load(opened)
+    return CrdsCfgManager(context, refpath, crdscfg)
 
 class CrdsCfgManager(object):
     """The CrdsCfgManager handles find."""
@@ -336,4 +179,4 @@ def test():
 # --------------------------------------------------------------------------------------
 
 if __name__ == "__main__":
-    print(generate_crdscfg_yaml())
+    print(test())
